% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgal4h-package.R
\docType{package}
\name{cgal4h-package}
\alias{cgal4h}
\alias{cgal4h-package}
\title{Provides the 'C++' header-only library of CGAL version 4}
\description{
'CGAL' is a C++ library that aims to provide easy access to efficient and
  reliable algorithms in computational geometry. Since its version 4, 'CGAL' can be used
  as standalone header-only library and is available under a double GPL-3|LGPL license.
  <https://www.cgal.org/>.
}
\examples{
\dontrun{

# Include a dependency to cgal4h inside a cpp file
// [[Rcpp::depends(cgal4h)]]
include <Rcpp.h>
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/cgal4h}
  \item Report bugs at \url{https://gitlab.com/dickoa/cgal4h/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

}
